fix: merge float cannot show
